Rewrote to take advantage of PostSharp
If we are utilizing PostSharp diagnostics for logging, then we do not have to get the hierarchy repository using LoggerRepositoryManager.GetHierarchyRepository(); one is provided for us.  We just return the provided reopsitory, cast to log4net.Repository.Hierarchy.Hierarchy, and then we access the Root property of whatever object we get, and return that.